+++ python/common/business-registry-account/.env.sample
@@ -1,7 +1,10 @@
-AUTH_SVC_URL=https://authserviceurl
+AUTH_API_URL=https://auth-api-dev-142173140222.northamerica-northeast1.run.app
+AUTH_API_VERSION=/api/v1
+
 ACCOUNT_SVC_AUTH_URL=https://tokenUrl
 ACCOUNT_SVC_CLIENT_ID=client-id
 ACCOUNT_SVC_CLIENT_SECRET=secret-value
+ACCOUNT_SVC_TIMEOUT=300
 
 JWT_OIDC_WELL_KNOWN_CONFIG=well-known-config
 JWT_OIDC_ALGORITHMS=alg
@@ -14,5 +17,3 @@ JWT_OIDC_USERNAME=username
 JWT_OIDC_FIRSTNAME=firstname
 JWT_OIDC_LASTNAME=lastname
 JWT_OIDC_JWKS_CACHE_TIMEOUT=300
-
-ACCOUNT_SVC_TIMEOUT=300
