+++ python/common/business-registry-account/tests/conftest.py
@@ -12,14 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Common setup and fixtures for the pytest suite used by this service."""
-import json
-import os
 from contextlib import contextmanager
 
 import pytest
 from flask import Flask
 
-from business_account.config import CONFIGURATION
+from business_account import AccountService
+from business_account.config import TestConfig
 from flask_jwt_oidc import JwtManager
 
 _jwt = JwtManager()
@@ -36,7 +35,7 @@ def not_raises(exception):
         raise pytest.fail(f"DID RAISE {exception}")
 
 
-def setup_jwt_manager(app, jwt_manager):
+def setup_jwt_manager(app: Flask, jwt_manager: JwtManager):
     """Use flask app to configure the JWTManager to work for a particular Realm."""
     def get_roles(a_dict):
         return a_dict["realm_access"]["roles"]  # pragma: no cover
@@ -45,10 +44,10 @@ def setup_jwt_manager(app, jwt_manager):
     jwt_manager.init_app(app)
 
 
-def create_app(run_mode=os.getenv("FLASK_ENV", "production")):
+def create_app():
     """Return a configured Flask App using the Factory method."""
     app = Flask(__name__)
-    app.config.from_object(CONFIGURATION[run_mode])
+    app.config.from_object(TestConfig)
     setup_jwt_manager(app, _jwt)
 
     return app
@@ -56,38 +55,14 @@ def create_app(run_mode=os.getenv("FLASK_ENV", "production")):
 @pytest.fixture(scope="session")
 def app():
     """Return a session-wide application configured in TEST mode."""
-    _app = create_app("testing")
+    _app = create_app()
 
     return _app
 
-
 @pytest.fixture(scope="function")
-def app_ctx(event_loop):
-    # def app_ctx():
-    """Return a session-wide application configured in TEST mode."""
-    _app = create_app("testing")
-    with _app.app_context():
-        yield _app
-
-
-@pytest.fixture
-def config(app):
-    """Return the application config."""
-    return app.config
-
-
-@pytest.fixture(scope="function")
-def app_request():
-    """Return a session-wide application configured in TEST mode."""
-    _app = create_app("testing")
-
-    return _app
-
-
-@pytest.fixture(scope="session")
-def client(app):  # pylint: disable=redefined-outer-name
-    """Return a session-wide Flask test client."""
-    return app.test_client()
+def account_service(app):
+    with app.app_context():
+        yield AccountService()
 
 
 @pytest.fixture(scope="session")
